Initial commit with progress made today.
On a total whim today, I learned the basics of Love2d and immediately became productive in it. Also on a whim, I decided to see how fast I could get a version of my old HyperChromatic Terrain project running in it. I have it working now, but I implemented it very hastily and sloppily.

Committing now because I will begin to refactor in a dramatic way soon.