Change settings to that it is runnable in web browser
I returned to this on a whim to see if it was possible to export it to be runnable in a web browser via the Love Web Builder and love.js. At first, it wouldn't run, but after reading the info on that page, realized chaning the dimensions via setMode was the problem. After undoing that, it became runnable, but of course it runs what seems 100x slower in-browser, due to non-JIT performance.

Well, I now understand the process of exporting Love2d / Lua games to the web, and thanks to these tools, it is extremely trivial.

Next up -- well, anything really. This application was ported on a whim, and there are dozens upon dozens of improvements that it would make it better - resolution controllers, scale controllers, toggles for the many settings like octaves, lacunarity, persistence, velocity for each of them, number of colors, images as colors, sounds, music, built-in help, an introduction, a logo... really, truly anything. It would also be worthwhile to implement much much better underlying data structures and algorithms, and perhaps to use GLSL / WebGL shaders if possible (this would be truly immense in terms of speed up). I'm also curious to see just how much Fennel can be used here. I know it can be used, but just what percentage of the project can be Fennel? Should work to find out.