Port noise-generator from Lua to Fennel
Started filling out the core modules, and then focused on porting over the noise generating functions. Nothing has been tested yet, and I won't test anything until everything is ported (maybe). The idea is to port everything, and then once we have something working, iteratively refactor until it takes the exact shape we want, and until it is in an easily extensible form.

These noise functions will definitely need to be refactored to be more concise and readable, but that will come later.

Next up is to continue porting over other functions, or to try to improve what was already ported over.