Port over more functions, focusing on height palette array generation
I'm just trying to get a port going, bad code and all. Once everything is ported and running, then I'll focus on making it better and adhering to a functional paradigm. Just want to get it done right now. swap! swap-if! and get-random-array tested and working right now.

Next up is to continue porting functions.