Start refactoring height resequencing to subtract 1.0
Started implementing the part of the increment function that reduces heights greater that 1.0 by 1.0 to loop them to the bottom of the palette, but realized it really had to be added to the resequencing function to work in the current implementation. It doesn't really make logical sense for it to be there, but I don't want to spend too much time making things make sense until I have the program operational, at which point I can begin substantive changes to the program.

I'm running into a bug whereby I'm eithering getting proper incrementing but not looping, or proper looping but not incrementing - I obviously need both. I'm sure its trivial but I'm not seeing it at a glance.

Committing so I can hack away recklessly and without concern on the program as I chase this bug.

Next up is to rectify the bug preventing height incrementation from working properly.