Resolve height-incrementation bug, completing function increment-heights!
The bug was exactly where I had thought it would be located, and the issue was that a 'temp' reference be utilized so that a needed value isn't overwritten in a potential swap. Tested to work via REPL, but not comprehensively.

This is definitely a place where unit tests and integration tests would be valuable. Noted for the future when things get serious.

Next up is to continue porting over and 'lightly' refactoring whatever functions remain from the Lua version of the program.