Write convenience script for launching 'game'; begin writing core loop
I'm at the point where the remaining functions in the original version of the program are only related to Love2d. My first version was quick'n'dirty - I'm now trying to write it in a better and more scalable manner.

Due to the fact that I want to write in a mostly functional style, I will have to override Love's functions to replace them with ones that utilize globals less to ones that pass state around.

I'm currently reading up on the default implementations, to see exactly what the engine expects to exist and in what manner. My understanding is that the system is entirely customizable and entirely overridable, but I need to see the exact limits.

Currently runs, but only displays a flashing blue rectangle (just trivial graphics to confirm it runs). I was planning on doing TCO for love.run but am not sure it really adds much to the loop, and it may be better to just use a loop over it, with the core loop returning whether or not it is to continue as part of its returning state.

Next up is to continue implementing the core functions of the game loop. There is also a lot of thinking that needs to go into exactly what is going to comprise 'state', and what the various components of state will need to be.