Write conf.lua to explicitly state core settings
I'm still reading over the default implementation of things that Love2d implements in the absence of its functions being overwritten, and I learned of conf.lua and decided to make one to lock in the core settings for now.

I've also been thinking hard about how I want to personally implement the core loop, and had a certified Buddha moment and have been whigging out hard for the last 30 minutes - I already understand monads, but now I really GET it. I was thinking about how if I allowed a side effect in any function deep into my program, that technicially every function above it also became impure. Thus, even a single impure function can render an entire module or program impure. I was thinking about how to resolve this, since I only wanted side effects to be at the input processing function, and it occurred to me to have a side-effect request system that punts the processing of the side-effect to the input handling section, when I realized that I was describing was just Haskell's monads!

Now I've worked with monads before and already understood the tree-like structure of a strictly functional program, where the IO monad forces there to be a impure/pure membrane, so that devs can clearly tell when they are in impure territory. But now I truly feel it in my bones, with every fiber of my being. Now that this 'punting' mechanism has been internalized, I can see it rippling out into so many of my programs. What an amazing feeling. I will never look the same at implementing programs again. From my experience, this isn't the last time I will feel this way.

Next up is to continue thinking about and implementing the core loop, and the state and substates that will be needed to facilitate. I've also realized that I will need an update state function that takes a keyword and update, so that the whole state can be shallow copied where possible and deep-mutated where necessary, instead of deep copying it every frame. So that may be something I want to skeleton out soon.