Outline desired functional program structure with placeholder functions
I thought quite a bit on exactly how I want to execute this, and what needs to be overridden from the Love2d engine. This is the first step in implementing that.

I opted for TCO in the main loop via the continue-or-quit! function, and by only allow side-effects from a limited subset of functions callable from loop!.

Next up is to work towards a minimal program that first renders just anything basic at all (like a square or circle) as a token intermediate step, and then the actual rendering function that I had before in the Lua version of the program.