Render a trivial figure in the newly implemented engine loop
Got the ol' blue rectangle blinking once more. Proof-of-concept fully established - this engine is ready to go full-steam ahead now!

I spent a bit of time thinking about how to do this - I know my goal is to get the GPU doing the rendering asap, but I still feel like I need to port my CPU-based code from the Lua implementation. However, there are several ways to do it, but all of them are at odds with functional programming. I can accumulate a list of draw operations as thunks, and then actually execute the thunks in the side-effect function render-to-screen!, or I can simply store a table of parameters for each draw operation and then use a case/match statement to call the appropriate functions in the side-effect function. Both of these require a fair amount of overhead for storing tables or thunks, but they would in fact adhere to functional paradigms. The less computationally and memory intensive way would be to simply render them on the spot as they are calculated, but then this moves side-effects into areas of the loop not meant for such things. And ultimately, the GPU way of doing things is naturally functional - I merely store the shader and then the shader does everything at once! I'll thus be implementing something that I'll be undoing later on anyway. I think I've settled on storing an array of pixels or pixel batches in process-graphics, and then actually blitting them in render-to-screen!, as a good middle ground that isn't too taxing in resources or time, and isn't too convoluted.

Next up is to continue implementing the needed functions to get a CPU-based render system going.